# artifacts/openapi.yaml
openapi: 3.0.3
info:
  title: EasyDocument API
  version: 1.0.0
  description: API contract for the EasyDocument Nepal document-agent platform.
servers:
  - url: https://api.easydocument.local/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SendOtpRequest:
      type: object
      required: [phoneNumber, purpose]
      properties:
        phoneNumber: { type: string, example: "+9779800000000" }
        purpose: { type: string, enum: [REGISTER, LOGIN, RESET] }
    VerifyOtpRequest:
      type: object
      required: [phoneNumber, otp]
      properties:
        phoneNumber: { type: string }
        otp: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          type: object
...
paths:
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendOtpRequest' }
      responses:
        '200': { description: OTP sent }
        '429': { description: Rate limited }
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and create session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
...